refactor(frontend): extract nativeChange test helper to shared util
